requirements.txt
streamlit
openai
langchain
faiss-cpu
pdfplumber
python-docx
docx2txt
spacy
fuzzywuzzy
python-Levenshtein
pandas
sqlalchemy
tqdm
python-dotenv
requests
